disable delete of unsaved card

+++ app/components/card-item.js
@@ -4,6 +4,9 @@ export default Ember.Component.extend({
   tagName: 'li',
   card: null,
 
+  disableDelete: Ember.computed('card.isNew', function(){
+    return this.get('card.isNew');
+  }),
   click: function(){
     let card = this.get('card');
     if (!card.get('belongsToCurrentUser')){
@@ -13,7 +16,11 @@ export default Ember.Component.extend({
   },
   actions: {
     deleteCard(card) {
-      card.destroyRecord();
+      //seting isDisabled to false on delete btn
+      //doesn't disable this event; seems bug in ember-cli-materialize
+      if (!this.get('disableDelete')){
+        card.destroyRecord();
+      }
     },
     saveCard(card) {
       this.sendAction('saveCard', card);
